Services must execute the completion while polling
This allows using the services directly.
If a completion throws an exception the service must ensure to stay in a valid state
Completions must be called on the same thread which called `.poll()`
The `tcx::async_*` functions post the completions to an executor, this allows using multiple threads in executors.